requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.6.0
    # via pydantic
anthropic==0.30.0
    # via langchain-anthropic
anyio==3.7.1
    # via
    #   anthropic
    #   httpx
    #   langchain-core
    #   openai
arrow==1.3.0
    # via -r requirements/requirements.in
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via -r requirements/requirements.in
defusedxml==0.7.1
    # via langchain-anthropic
distro==1.8.0
    # via
    #   anthropic
    #   openai
filelock==3.15.4
    # via huggingface-hub
fsspec==2024.6.1
    # via huggingface-hub
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via -r requirements/requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.1
    # via
    #   anthropic
    #   openai
huggingface-hub==0.23.4
    # via tokenizers
humanize==4.8.0
    # via -r requirements/requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.5.0
    # via anthropic
jsonpatch==1.33
    # via langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain-anthropic==0.1.4
    # via -r requirements/requirements.in
langchain-core==0.1.12
    # via
    #   -r requirements/requirements.in
    #   langchain-anthropic
    #   langchain-openai
langchain-openai==0.0.2.post1
    # via -r requirements/requirements.in
langsmith==0.0.65
    # via langchain-core
loguru==0.7.2
    # via -r requirements/requirements.in
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.2
    # via langchain-openai
openai==1.35.7
    # via langchain-openai
packaging==23.2
    # via
    #   huggingface-hub
    #   langchain-core
prompt-toolkit==3.0.41
    # via -r requirements/requirements.in
pydantic==2.5.1
    # via
    #   anthropic
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.14.3
    # via pydantic
pygments==2.17.1
    # via rich
pyperclip==1.8.2
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via arrow
pyyaml==6.0.1
    # via
    #   -r requirements/requirements.in
    #   huggingface-hub
    #   langchain-core
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   tiktoken
rich==13.7.0
    # via -r requirements/requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.0
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
tenacity==8.2.3
    # via langchain-core
tiktoken==0.5.2
    # via
    #   -r requirements/requirements.in
    #   langchain-openai
tokenizers==0.19.1
    # via anthropic
tqdm==4.66.1
    # via
    #   huggingface-hub
    #   openai
types-python-dateutil==2.8.19.14
    # via arrow
typing-extensions==4.8.0
    # via
    #   anthropic
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.1.0
    # via requests
wcwidth==0.2.10
    # via prompt-toolkit
